split long and wide
